Renamed service to utility
