+++ FolderInspector/FolderInspector.csproj
@@ -59,9 +59,9 @@
     <Compile Include="Constants\AppSettings.cs" />
     <Compile Include="FolderInspector.cs" />
     <Compile Include="Properties\AssemblyInfo.cs" />
-    <Compile Include="Services\ExcelService.cs" />
-    <Compile Include="Services\FolderService.cs" />
-    <Compile Include="Services\WordService.cs" />
+    <Compile Include="Utilities\ExcelUtility.cs" />
+    <Compile Include="Utilities\FolderUtility.cs" />
+    <Compile Include="Utilities\WordUtility.cs" />
   </ItemGroup>
   <ItemGroup>
     <None Include="App.config" />
